Began work on Skillset and saving
Skillset is a simple class used to manage the (currently) four values
that are the players skills.  Savable is an interface that requires
implementing objects to return a JSONObject, though that doesn't yet
work due to some struggles with my classpath.  Database will eventually
take in savable objects and write them to files, and help read save
files back into game objects.